LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/text/2d.text.align.default.worker-expected.txt


FAIL 2d assert_equals: ctx.textAlign === 'start' (got [undefined], expected start[string]) expected (string) "start" but got (undefined) undefined
